refactor(6776): clarify kolme docs support module naming
