+++ crates/kamn-core/tests/kolme_devnet_ops_docs/local_lane_contract_tests/local_heavy_lane_contract_tests.rs
@@ -1,4 +1,4 @@
-use super::super::shared_support::{assert_plan_contains_all};
+use super::super::docs_assert_support::{assert_plan_contains_all};
 
 const PLAN_CONTAINS_LOCAL_ONLY_HEAVY_VALIDATION_MATRIX_PLAN_MARKERS: &[&str] = &[
     "## Local-Only Heavy Kolme Validation Matrix",
