+++ crates/kamn-core/tests/kolme_devnet_ops_docs/regression_local_lane_contract_tests/local_fork_regression_contract_tests.rs
@@ -1,4 +1,4 @@
-use super::super::shared_support::{assert_plan_contains_all};
+use super::super::docs_assert_support::{assert_plan_contains_all};
 
 const REGRESSION_REQUIRES_LOCAL_FORK_SYNC_METADATA_GUARD_MARKER_PLAN_MARKERS: &[&str] = &[
     "local fork metadata sync lane fails closed for checkout-path, remote-URL, ref, and dirty-checkout drift (`Regression: #1429`).",
