+++ crates/kamn-core/tests/kolme_devnet_ops_docs/runtime_transport_contract_tests.rs
@@ -1,4 +1,4 @@
-use super::shared_support::{assert_plan_contains_all};
+use super::docs_assert_support::{assert_plan_contains_all};
 
 const PLAN_CONTAINS_RUNTIME_TRANSPORT_RETRY_RECONNECT_FAILURE_TAXONOMY_PLAN_MARKERS: &[&str] = &[
     "## Runtime Transport Retry-Reconnect Failure Taxonomy (Issue #4508)",
